requirements.txt
appnope==0.1.3
asttokens==2.4.0
backcall==0.2.0
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
executing==2.0.0
ipykernel==6.26.0
ipython==8.16.1
iterative-stratification==0.1.7
jedi==0.19.1
joblib==1.3.2
jupyter_client==8.5.0
jupyter_core==5.4.0
lingua-language-detector==2.0.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.8
numpy==1.24.4
packaging==23.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
polars==0.19.11
prompt-toolkit==3.0.39
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.16.1
python-dateutil==2.8.2
pyzmq==25.1.1
regex==2023.10.3
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
stack-data==0.6.3
threadpoolctl==3.2.0
tornado==6.3.3
traitlets==5.12.0
wcwidth==0.2.8
xlsx2csv==0.8.1
